Klibs are now built with build.go. Just ./BUILD klibs as usual, no need to clean before. Begin migrating libs, for now libc/libip/libdraw

+++ sys/src/libc/port/pool.c
@@ -475,8 +475,8 @@ blockmerge(Pool *pool, Bhdr *a, Bhdr *b)
 
 	t = B2T(a);
 	t->size = (uint32_t)Poison;
-	t->magic0 = NOT_MAGIC;
-	t->magic1 = NOT_MAGIC;
+	t->magic0 = (uint8_t)NOT_MAGIC;
+	t->magic1 = (uint8_t)NOT_MAGIC;
 	PSHORT(t->datasize, NOT_MAGIC);
 
 	a->size += b->size;
